pyproject.toml
[tool.poetry]
name = "callsign-alert"
version = "0.1.0"
description = "A tool to alert users when their desired amateur radio callsign is available."
authors = ["Amin Shah Gilani <amin@gilani.me>"]
license = "AGPL-3.0-only"

[tool.poetry.dependencies]
python = "^3.11"
Django = "3.2"
django-encrypted-secrets = "^0.9.9"

[tool.poetry.dev-dependencies]
black = "22.12.0"
isort = "^5.11.4"
wemake-python-styleguide = "^0.17.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
